% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ngraph_shiny.R
\name{Ngraph}
\alias{Ngraph}
\title{Interactive Plotting with Shiny}
\usage{
Ngraph(data, variable, graph)
}
\arguments{
\item{data}{If the inputted data contains more than one cycle, this function generates a graph over time}

\item{variable}{User can specify the variables to plot with this function.}

\item{graph}{Type of graph the user wishes to generate}
}
\description{
Interactive Plotting with Shiny
}
