% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ngraph_table.R
\name{Ngraph_table}
\alias{Ngraph_table}
\title{Title}
\usage{
Ngraph_table(year, variables)
}
\arguments{
\item{year}{String vector that contains a range of years or a specific year}

\item{variables}{}
}
\description{
Title
}
