% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_file_list.R
\docType{data}
\name{nhanes_file_list}
\alias{nhanes_file_list}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 1713 rows and 10 columns.
}
\usage{
nhanes_file_list
}
\description{
This is data to be included in my package
}
\references{
\url{https://cran.r-project.org/web/packages/RNHANES/vignettes/introduction.html}
}
\author{
Minjee Kim \email{minjeekim0804@tamu.edu}
}
\keyword{datasets}
