% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_variable_list.R
\docType{data}
\name{nhanes_variable_list}
\alias{nhanes_variable_list}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 60383 rows and 10 columns.
}
\usage{
nhanes_variable_list
}
\description{
This is data to be included in my package
}
\references{
\url{https://cran.r-project.org/web/packages/RNHANES/vignettes/introduction.html}
}
\author{
Herb Susmann
}
\keyword{datasets}
