% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_viz.R
\name{viz}
\alias{viz}
\title{Visualize variables}
\usage{
viz(type, dat, var)
}
\arguments{
\item{type}{type of data, inherited from nhanes_viz}

\item{dat}{data, filtered using file_name and cycle}

\item{var}{variable name, inherited from nhanes_viz}
}
\value{
ggplot graph
}
\description{
Visualize variables
}
