% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ngraph_table.R
\name{year_check}
\alias{year_check}
\title{Year check}
\usage{
year_check(first, second)
}
\arguments{
\item{first}{beginning of the range}

\item{second}{end of the range}
}
\value{
all cycles between the inputted range of years
}
\description{
Year check
}
