% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nGraph_table.R
\name{nhanes_table}
\alias{nhanes_table}
\title{Loading NHANES data with a flexible year input}
\usage{
nhanes_table(year = NULL, file_name = NULL, demographics = F, recode = F)
}
\arguments{
\item{year}{a numeric value of a year or string objective of a range year}

\item{file_name}{name of the file that the user wishes to load (required)}

\item{demographics}{logical value - whether to include a demographic information (inherited from RNHANES::nhanes_search())}

\item{recode}{logical value - whether to recode the categorical variables into integers (inherited from RNHANES::nhanes_search())}
}
\value{
an NHANES data table with user specified parameters
}
\description{
Loading NHANES data with a flexible year input
}
\examples{
\dontrun{
nhanes_table("2009","EPH")
}
}
