% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ngraph_table.R
\name{nhanes_table}
\alias{nhanes_table}
\title{Create a data frame with specified parameters}
\usage{
nhanes_table(year, variables)
}
\arguments{
\item{year}{String vector that contains a range of years or a specific year}

\item{variables}{A list or a specific variable the user is interested in. Search for a specific variable using the "nhanesA::nhanesTableVars('EXAM', 'BMX_D')"}
}
\description{
Create a data frame with specified parameters
}
