% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesGraph-package.R
\docType{package}
\name{nhanesGraph-package}
\alias{nhanesGraph}
\alias{nhanesGraph-package}
\title{nhanesGraph: Graphical tools for NHANES data}
\description{
Graphical data search tool for the US National Health and Nutrition Examination Survey data.
}
\author{
\strong{Maintainer}: Minjee Kim \email{minjeekim0804@tamu.edu} (\href{https://orcid.org/0000-0002-2001-2442}{ORCID})

}
\keyword{internal}
